+++ docs/release-notes.md
@@ -9,6 +9,8 @@
 * [1243672](https://bugzilla.redhat.com/show_bug.cgi?id=1243672) - Source refs does not shows up in editor
 * [1062476](https://bugzilla.redhat.com/show_bug.cgi?id=1062476) - User can attempt to upload an unspecified glossary file
 * [1243251](https://bugzilla.redhat.com/show_bug.cgi?id=1243251) - Project search does not sort correctly by date over multiple pages
+* [1243688](https://bugzilla.redhat.com/show_bug.cgi?id=1243688) - Groups filter by archived not working
+
 
 -----------------------